free disk space calculation fixed for 32bit system

+++ util/util/path/status.cpp
@@ -168,7 +168,8 @@ util::Error FreeDiskSpace(const std::string& real, unsigned long long& freeBytes
   if (statvfs(real.c_str(), &sfs) <0)
     return util::Error::Failure(errno);
 
-  freeBytes = sfs.f_bsize * sfs.f_bavail;
+  freeBytes = static_cast<unsigned long long>(sfs.f_bsize) * 
+              static_cast<unsigned long long>(sfs.f_bavail);
   return util::Error::Success();
 }
